Match func_80023D08 in C
Use a signed 16-bit glide quotient so GCC naturally keeps both division results in v0 and assigns the cursor-delay constants to v1. This removes the candidate's hard register pin while preserving the exact executable.

+++ src/game/func_80023D08.c
@@ -15,7 +15,7 @@ void func_80023D08(GridCursor *o, s32 dir) {
     s32 pos;
     s32 shift;
     s32 index;
-    s32 q;
+    s16 q;
 
     d = o->object;
     if (o->flags & 0x80) {
